new file:   autumn/return-date-number/rb/b.rb

+++ autumn/return-date-number/rb/b.rb
@@ -1,3 +1,3 @@
 o = Time.now
-n = o.to_f
+n = o.to_i
 puts n
